spread test to registration and ubuntuproinfo
